New features

+++ KiloMart.Core/Models/PaginatedResult.cs
@@ -5,5 +5,5 @@ public class PaginatedResult<T>
     public int TotalCount { get; set; }
     public int PageNumber { get; set; }
     public int PageSize { get; set; }
-    public int TotalPages => (int)Math.Ceiling((double)TotalCount / PageSize);
+    public int TotalPages => (int)Math.Ceiling((decimal)TotalCount / PageSize);
 }
